requirements.txt
tensorflow_cpu==2.6.0
streamlit==0.88.0
transformers==4.10.2
requests==2.26.0
numpy==1.19.5
pandas==1.3.3
altair==4.1.0
tensorflow_hub==0.12.0
Pillow==8.3.2
